Remove WinitWindows, use a single window.
