+++ engine/lib/phx/src/window/mod.rs
@@ -10,7 +10,6 @@ mod window_resize_constraints;
 mod window_resolution;
 mod winit_converters;
 mod winit_window;
-mod winit_windows;
 
 pub use composite_alpha_mode::*;
 pub use cursor::*;
@@ -24,7 +23,6 @@ pub use window_resize_constraints::*;
 pub use window_resolution::*;
 pub use winit_converters::*;
 pub use winit_window::*;
-pub use winit_windows::*;
 
 use internal::ConvertIntoString;
